+++ packages/orq-rc/RELEASES.md
@@ -12279,3 +12279,13 @@ Based on:
 - [typescript v4.15.0-rc.39] packages/orq-rc
 ### Releases
 - [NPM v4.15.0-rc.39] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.39 - packages/orq-rc
+
+## 2026-09-11 12:18:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.4 (2.935.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.15.0-rc.40] packages/orq-rc
+### Releases
+- [NPM v4.15.0-rc.40] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.40 - packages/orq-rc